Allocate arrays on heap and add named type definitions
Dynamically allocate boxed arrays on heap instead of stack. This can be
optimized later to use the stack for local, constant allocations. Add
named type definitions and type checking. Replace TStruct with a TNamed
that contains a StructDef. Refactor type checking and annotation phase
to use two type environments, one for values and one for types.